♻️ refactor(custom): 관심사 응집도를 위한 폴더 구조 변경
